🛠️ Improve and ⚡ Optimize

+++ backend/models/User.js
@@ -12,7 +12,6 @@ const userSchema = new mongoose.Schema(
       unique: true, // Ensures username is unique
       minlength: 3,
       maxlength: 30,
-      required: true,
       lowercase: true,
     },
     email: {
@@ -70,12 +69,12 @@ const userSchema = new mongoose.Schema(
         enum: ["pending", "approved", "rejected"],
         default: "pending",
       },
-      rating: {
-        type: Number,
-        min: 1,
-        max: 5,
-      },
       documents: { type: [String], default: [] },
+    },
+    rating: {
+      type: Number,
+      min: 1,
+      max: 5,
     },
   },
   { timestamps: true } // Automatically adds `createdAt` and `updatedAt`
@@ -89,7 +88,10 @@ userSchema.pre("save", async function (next) {
 });
 
 userSchema.post("findOneAndUpdate", async function (doc) {
-  if (doc.workerApplication.status === "rejected") {
+  if (
+    doc.workerApplication.status === "rejected" &&
+    doc.workerApplication.documents?.length
+  ) {
     await cloudinary.api.delete_resources(
       doc.workerApplication.documents.map((d) => d.public_id)
     );
@@ -99,6 +101,11 @@ userSchema.post("findOneAndUpdate", async function (doc) {
 // Method to compare password
 userSchema.methods.validatePassword = async function (candidatePassword) {
   return await bcrypt.compare(candidatePassword, this.password); // Compare the input password with the hashed password
+};
+
+userSchema.methods.incrementTokenVersion = function () {
+  this.tokenVersion += 1;
+  return this.save();
 };
 
 const User = mongoose.model("User", userSchema);
